split sequence generation from database into new module
