+++ oligo_designer_toolsuite/database/_oligos_database.py
@@ -5,7 +5,6 @@
 import os
 import yaml
 import warnings
-import random
 import pandas as pd
 
 from pathlib import Path
@@ -16,13 +15,9 @@ from Bio import SeqIO
 from Bio.Seq import Seq
 from Bio.SeqRecord import SeqRecord
 
-from ..utils._data_parser import (
-    check_fasta_format,
-    check_tsv_format,
-    parse_fasta_header,
-)
-
-from ..utils._database_processor import collapse_info_for_duplicated_sequences, merge_databases
+from ..utils._utils import check_if_list, check_tsv_format
+from ..utils._sequence_parser import FastaParser
+from ..utils._database_processor import merge_databases, collapse_info_for_duplicated_sequences
 
 
 ############################################
@@ -73,6 +68,8 @@ class OligoDatabase:
         self.dir_output = os.path.abspath(os.path.join(dir_output, "oligo_database"))
         Path(self.dir_output).mkdir(parents=True, exist_ok=True)
 
+        self.fatsa_parser = FastaParser()
+
         # Initialize databse object
         self.database = {}
         self.oligosets = {}  # will be used later in the gereration of non overlpping sets
@@ -119,7 +116,7 @@ class OligoDatabase:
         :param file_database: File containing oligo database.
         :type file_database: str
         """
-        region_ids = self._check_if_list(region_ids)
+        region_ids = check_if_list(region_ids)
 
         if database_overwrite:
             warnings.warn("Overwriting database!")
@@ -169,16 +166,16 @@ class OligoDatabase:
         extend the database, not overwrite it.
         check if sequnce should be target or oligo, i.e. sequence type."""
 
-        region_ids = self._check_if_list(region_ids)
+        region_ids = check_if_list(region_ids)
 
         if database_overwrite:
             warnings.warn("Overwriting database!")
 
-        fasta_sequences = self._read_sequences(file_fasta_in, region_ids)
+        fasta_sequences = self.fatsa_parser.read_fasta_sequences(file_fasta_in, region_ids)
 
         region_sequences = {}
         for entry in fasta_sequences:
-            region, additional_info, coordinates = parse_fasta_header(entry.id)
+            region, additional_info, coordinates = self.fatsa_parser.parse_fasta_header(entry.id)
             oligo_info = coordinates | additional_info
             if region in region_sequences:
                 if entry.seq in region_sequences[region]:
@@ -208,123 +205,6 @@ class OligoDatabase:
             self._check_if_region_in_database(database_tmp, region_ids)
 
         self.database = database_tmp
-
-    def create_sequences_sliding_window(
-        self, filename_out: str, file_fasta_in: list[str], sequence_length: int, region_ids: list[str] = None
-    ):
-        """create new fasta files with tiled sequences from input sequences via sliding window.
-        Load sequences from input fasta file, and run a sliding window over them to create
-        tiled sequences with desired length.
-        save the generated sequences in fasta file with header information taken from input fasta
-        file but adjust coordinates."""
-
-        def get_sliding_window_sequence(entry, sequence_length, handle_fasta):
-            entry_sequence = entry.seq
-            region, additional_info, coordinates = parse_fasta_header(
-                header=entry.id, parse_additional_info=False
-            )
-            # chromosome and strand information is the same for an entry but parsed as
-            # list in the parse_fasta_header, hence, we take only the first element for each
-            # if no information provided in fasta file, those entries are None
-            chromosome = coordinates["chromosome"][0]
-            strand = coordinates["strand"][0]
-            list_of_coordinates = []
-            # if the fasta header DOES NOT contain coordinates information
-            if chromosome is None:
-                list_of_coordinates = [None for i in range(len(seq))]
-            # if the fasta header DOES contain coordinates information
-            else:
-                # coordinates in fasta file use 1-base indixing, which go from 1 (for base 1) to n (for base n)
-                # range produces values until end-1 (e.g. range(10) goes until 9) -> add +1
-                # the start and end coordinates can be a list for regions spanning split sequences (e.g. exon junctions, UTRs)
-                for i in range(len(coordinates["start"])):
-                    list_of_coordinates.extend(
-                        list(
-                            range(
-                                coordinates["start"][i],
-                                coordinates["end"][i] + 1,
-                            )
-                        )
-                    )
-
-            # sort reverse on minus strand becaus the sequence is translated into the reverse complement by fasta -strand option
-            if strand == "-":
-                list_of_coordinates.reverse()
-
-            # generate sequences with sliding window and write to fasta file (use lock to ensure that hat only one process can write to the file at any given time)
-            if len(entry_sequence) > sequence_length:
-                num_sequences = len(entry_sequence) - (sequence_length - 1)
-                sequences = [entry_sequence[i : i + sequence_length] for i in range(num_sequences)]
-
-                for i in range(num_sequences):
-                    seq = sequences[i]
-                    seq_start_end = [
-                        list_of_coordinates[i],
-                        list_of_coordinates[(i + sequence_length - 1)],
-                    ]
-                    # sort coordinates for oligos on minus strand
-                    start_seq = min(seq_start_end)  # 1-base index
-                    end_seq = max(seq_start_end)
-
-                    header = f"{region}::{additional_info}::{chromosome}:{start_seq}-{end_seq}({strand})"
-                    handle_fasta.write(f">{header}\n{seq}\n")
-
-        region_ids = self._check_if_list(region_ids)
-        file_fasta_in = self._check_if_list(file_fasta_in)
-
-        file_fasta_out = os.path.join(self.dir_output, f"{filename_out}.fna")
-        with open(file_fasta_out, "w") as handle_fasta:
-            for file_fasta in file_fasta_in:
-                fasta_sequences = self._read_sequences(file_fasta, region_ids)
-                # did not parallize this function because workers would be writing simultaneously to the same file
-                # which could lead to data corruption. I tried using the file_lock=Lock() function from multiprocessing
-                # as input to get_sliding_window_sequence(file_lock) and then use with file_lock: before writing to the file
-                # but this lead to an pickle error from joblib which I could not solve, so I left it like that for now
-                for entry in fasta_sequences:
-                    get_sliding_window_sequence(entry, sequence_length, handle_fasta)
-
-        return file_fasta_out
-
-    def create_sequences_random(
-        self,
-        filename_out: str,
-        length_sequences: int,
-        num_sequences: int,
-        name_sequences: str = "randomsequence",
-        base_alphabet_with_probability: list[float] = {"A": 0.25, "C": 0.25, "G": 0.25, "T": 0.25},
-    ):
-        """create sequences from random sequences with desired length and sequence composition.
-        save as fasta file, with appropriate header, e.g. region id and type=random_sequence."""
-
-        def get_sequence_random(sequence_length, base_alphabet_with_probability):
-            bases = list(base_alphabet_with_probability.keys())
-            sequence = "".join(
-                random.choices(
-                    bases, weights=[base_alphabet_with_probability[n] for n in bases], k=sequence_length
-                )
-            )
-            return sequence
-
-        # make sure to have n unique sequences
-        sequences_list = []
-        while len(sequences_list) < num_sequences:
-            num_missing_sequences = num_sequences - len(sequences_list)
-            missing_sequences = list(
-                set(
-                    [
-                        get_sequence_random(length_sequences, base_alphabet_with_probability)
-                        for i in range(num_missing_sequences)
-                    ]
-                )
-            )
-            sequences_list = list(set(sequences_list + missing_sequences))
-
-        file_fasta_out = os.path.join(self.dir_output, f"{filename_out}.fna")
-
-        with open(file_fasta_out, "w") as handle_fasta:
-            for seq in sequences_list:
-                handle_fasta.write(f">{name_sequences}::regiontype=random_sequence\n{seq}\n")
-        return file_fasta_out
 
     def remove_regions_with_insufficient_oligos(self, pipeline_step: str):
         """Deletes from the ``oligo_DB`` the regions (e.g. genes) which have less than
@@ -398,7 +278,7 @@ class OligoDatabase:
         :rtype: str
         """
         if region_ids:
-            region_ids = self._check_if_list(region_ids)
+            region_ids = check_if_list(region_ids)
         else:
             region_ids = self.database.keys()
 
@@ -466,27 +346,6 @@ class OligoDatabase:
 
         return dir_oligosets
 
-    def _read_sequences(self, file_fasta_in, region_ids):
-        # check if files exist and are in correct format
-        if os.path.exists(file_fasta_in):
-            if not check_fasta_format(file_fasta_in):
-                raise ValueError("Fasta file has incorrect format!")
-        else:
-            raise ValueError("Fasta file does not exist!")
-        # remove undesired regions already at the beginning
-        if region_ids:
-            fasta_sequences = []
-            with open(file_fasta_in, "r") as handle:
-                for entry in SeqIO.parse(handle, "fasta"):
-                    region, _, _ = parse_fasta_header(entry.id, parse_additional_info=False)
-                    if region in region_ids:
-                        fasta_sequences.append(entry)
-        else:
-            with open(file_fasta_in, "r") as handle:
-                fasta_sequences = list(SeqIO.parse(handle, "fasta"))
-
-        return fasta_sequences
-
     def _check_if_region_in_database(self, database, region_ids):
         # check if all regions provided in region list exist
         # in the regions retreived from the fasta file
@@ -504,8 +363,3 @@ class OligoDatabase:
             if key not in region_ids:
                 database.pop(key)
         return database
-
-    def _check_if_list(self, obj):
-        if obj:
-            obj = [obj] if not isinstance(obj, list) else obj
-        return obj
